[[Chore]] Move scope-manager to external file
